refactor: split vignette artifact validator
