+++ tools/vignette_artifact_checks/__init__.py
@@ -1,0 +1,1 @@
+"""Concern-specific vignette artifact integration checks."""
