+++ tools/vignette_artifact_checks/common.py
@@ -1,0 +1,30 @@
+"""Shared low-level helpers for vignette artifact checks."""
+
+from __future__ import annotations
+
+import subprocess
+from pathlib import Path
+
+__all__ = ["run"]
+
+
+def run(
+    repo: Path,
+    *args: str,
+    check: bool = True,
+    env: dict[str, str] | None = None,
+) -> subprocess.CompletedProcess[str]:
+    result = subprocess.run(
+        args,
+        cwd=repo,
+        check=False,
+        text=True,
+        capture_output=True,
+        env=env,
+    )
+    if check and result.returncode != 0:
+        raise AssertionError(
+            f"Command failed ({result.returncode}): {' '.join(args)}\n"
+            f"stdout:\n{result.stdout}\nstderr:\n{result.stderr}"
+        )
+    return result
